## March 2026 Refactor & RAM Optimization
- Refactor: Move static helper functions in spectrum.c to file scope (C standard compliance, no logic change, improves maintainability and build stability).
- RAM optimization: Marked lookup tables and buffers as const where possible.
- Confirmed: No functional/logic changes; all builds pass and runtime behavior is identical.

+++ App/misc.h
@@ -377,7 +377,7 @@ extern volatile uint8_t      gVFOStateResumeCountdown_500ms;
     extern volatile bool     gScheduleFM;
 #endif
 extern uint8_t               gIsLocked;
-extern volatile uint8_t      boot_counter_10ms;
+extern volatile uint16_t     boot_counter_10ms;
 
 #ifdef ENABLE_FEAT_N7SIX
     extern bool                  gK5startup;
